Original PSR Magic Button
The run.bat file needs to be edited to change the path to where you are hosting the application. Once you have changed the path it should be as simple as double clicking to execute as long as you have the appropriate delegations on your machine to run .bat files.